rename: extract_features => extract_acoustic_features
